feat: Change inputs to likes on users profile to selects inputs and add exception to duplicates on register user

+++ app/Http/Controllers/Web/Admin/ProfileController.php
@@ -27,11 +27,37 @@ class ProfileController extends Controller
             if($slug){
                 $authenticated = request()->user();
 
+                function organizedLikes($item){
+                    return [
+                        'avatar' => $item->avatar,
+                        'bibliography' => $item->bibliography,
+                        'birthday' => \Carbon\Carbon::parse($item->birthday)->format('Y-m-d'),
+                        'city' => $item->city,
+                        'country' => $item->country,
+                        'created_at' => $item->created_at,
+                        'external_links' => $item->external_links,
+                        'gender' => $item->gender,
+                        'id' => $item->id,
+                        'is_active' => $item->is_active,
+                        'likes' =>  [
+                            'like' => $item->likes->where('category', 'like')->values()->toArray(),
+                            'unlike' => $item->likes->where('category', 'unlike')->values()->toArray()
+                        ],
+                        'name' => $item->name,
+                        'nickname' => $item->nickname,
+                        'permissions_keys' => $item->permissions_keys,
+                        'slug' => $item->slug,
+                        'state' => $item->state,
+                        'updated_at' => $item->updated_at
+                    ];
+                }
+                
                 if ($authenticated->permissions_keys->intersect(['administrator', 'dev'])->isNotEmpty()) {
-                    return User::where('slug', $slug)->firstOrFail();
+                    $user = User::where('slug', $slug)->firstOrFail();
+                    return organizedLikes($user);
                 }
 
-                return $authenticated;
+                return organizedLikes($authenticated);
             }
         }catch(\Throwable $e){
             return $this->provideException($e);
@@ -46,7 +72,6 @@ class ProfileController extends Controller
                 'nickname' => 'required',
                 'gender' => 'required',
                 'birthday' => 'required',
-                'email' => 'required',
                 'city' => 'required',
                 'state' => 'required', 
                 'country' => 'required',
@@ -56,7 +81,6 @@ class ProfileController extends Controller
                 'nickname.required' => "Apelido",
                 'gender.required' => "Gênero",
                 'birthday' => 'Data de nascimento',
-                'email' => 'Email',
                 'city' => 'Cidade',
                 'state' => 'Estado',
                 'country' => 'País',
@@ -71,7 +95,6 @@ class ProfileController extends Controller
                 'nickname' => $request->input('nickname', $user->nickname),
                 'gender' => $request->input('gender', $user->gender),
                 'birthday' => $request->input('birthday', $user->birthday),
-                'email' => $request->input('email', $user->email),
                 'city' => $request->input('city', $user->city),
                 'state' => $request->input('state', $user->state),
                 'country' => $request->input('country', $user->country),
@@ -80,34 +103,44 @@ class ProfileController extends Controller
 
             // Links externos
             $externalLinks = $request->input('external_links', []);
-            $externalLinksExisting = collect($externalLinks)->pluck('id')->filter()->toArray();
-            
-            UserExternalLink::where('user_id', $id)->whereNotIn('id', $externalLinksExisting)->delete();
-
-            foreach($externalLinks as $item){
-                if(isset($item['id'])){
-                    UserExternalLink::where('id', $item['id'])->update([
-                        'name' => $item['name'],
-                        'url' => $item['url'],
-                    ]);
-                }else{
-                    UserExternalLink::create([
-                        'user_id' => $id,
-                        'name' => $item['name'],
-                        'url' => $item['url'],
-                    ]);
+            if($externalLinks){
+                $externalLinksExisting = collect($externalLinks)->pluck('id')->filter()->toArray();
+                
+                UserExternalLink::where('user_id', $id)->whereNotIn('id', $externalLinksExisting)->delete();
+    
+                foreach($externalLinks as $item){
+                    if(isset($item['id'])){
+                        UserExternalLink::where('id', $item['id'])->update([
+                            'name' => $item['name'],
+                            'url' => $item['url'],
+                        ]);
+                    }else{
+                        UserExternalLink::create([
+                            'user_id' => $id,
+                            'name' => $item['name'],
+                            'url' => $item['url'],
+                        ]);
+                    }
                 }
             }
 
             // Gostos
             $likes = $request->input('likes', []);
-            foreach($likes as $item){
-                UserLike::where('id', $item['id'])->update([
-                    'category' => $item['category'],
-                    'content' => $item['content']
-                ]);
+            if($likes){
+                foreach($likes['like'] as $item){
+                    UserLike::where('id', $item['id'])->update([
+                        'category' => $item['category'],
+                        'content' => $item['content']
+                    ]);
+                }
+                foreach($likes['unlike'] as $item){
+                    UserLike::where('id', $item['id'])->update([
+                        'category' => $item['category'],
+                        'content' => $item['content']
+                    ]);
+                }
             }
-
+        
             return $this->provideSuccess('update');
         }catch(\Throwable $e){
             return $this->provideException($e);
